pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
sources = ["src"]

[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "busty"
version = "0.1.0"
description = "A Discord bot for playing music and managing media"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
license-files = ["LICENSE"]
keywords = ["discord", "bot", "music", "media"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "discord_emoji>=1.4.1",
    "mutagen>=1.47.0",
    "discord.py[voice] @ git+https://github.com/Rapptz/discord.py.git@cb7300990f656c0964ea48115354f9416e96dcd1",
    "PyNaCl>=1.5.0",
    "google-api-python-client>=2.57.0",
    "google-auth>=2.45.0",
    "google-auth-oauthlib>=1.2.0",
    "pillow>=12.0.0",
    "openai>=2.6.0",
    "tiktoken>=0.3.1",
    "types-requests>=2.31.0",
    "types-Pillow>=10.0.0",
    "colorlog>=6.8.0",
    "pyyaml>=6.0.0",
    "types-PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.1.0", "pyright>=1.1.0"]

[project.scripts]
busty = "busty.main:run_bot"

[dependency-groups]
dev = [
    "types-pillow>=10.2.0.20240822",
    "types-requests>=2.32.4.20250913",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"

[tool.mypy]
python_version = "3.13"
disallow_any_generics = true
disallow_incomplete_defs = true
check_untyped_defs = true
warn_return_any = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
warn_no_return = true
no_implicit_reexport = true
extra_checks = true

# These stay permissive for discord.py and other untyped dependencies
disallow_untyped_decorators = false
ignore_missing_imports = true
[[tool.mypy.overrides]]
module = "requests"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
